LotoPuntoApp/requirements.txt

Django==3.0.7
gunicorn==20.0.4
requests==2.24.0
psycopg2==2.8.5
djangorestframework==3.11.0